Move theme and global styles into theme folder, delete old app router
