--- genai-cookbook/app-orig/cookbook/image-captioning/api/route.ts
@@ -1,43 +1,0 @@
-import { generateText } from 'ai'
-import { ModelPreparationError, prepareModel } from '@/lib/prepareModel'
-
-/*
- * The captioning API mirrors our multi-turn chat route but returns a single
- * string instead of a streaming response. Because Modular MAX speaks the
- * OpenAI-compatible protocol, the Vercel AI SDK can works with Modular MAX
- * out of the box.
- */
-
-// ============================================================================
-// POST /api — generates an image caption
-// ============================================================================
-/** Processes caption requests for either Modular MAX or OpenAI. */
-export async function POST(req: Request) {
-    const { messages, endpointId, modelName } = await req.json()
-    if (!messages) {
-        return new Response('Client did not provide messages', { status: 400 })
-    }
-
-    let model
-    try {
-        model = await prepareModel(endpointId, modelName)
-    } catch (error) {
-        const modelError = error as ModelPreparationError
-        return new Response(modelError.message, { status: modelError.status })
-    }
-
-    try {
-        const { text } = await generateText({
-            // generateText handles the familiar chat-completions format via the Vercel AI SDK.
-            model: model,
-            messages: messages,
-        })
-
-        return Response.json({ text })
-    } catch (error) {
-        const errorMessage = error instanceof Error ? `(${error.message})` : ''
-        return new Response(`Failed to generate caption ${errorMessage}`, {
-            status: 424,
-        })
-    }
-}
