--- genai-cookbook/app-orig/cookbook/image-captioning/code/page.tsx
@@ -1,7 +1,0 @@
-'use client'
-
-import { RecipeCode } from '@/components/recipe-partials'
-
-export default function Code() {
-    return <RecipeCode />
-}
