--- genai-cookbook/app-orig/cookbook/image-captioning/layout.tsx
@@ -1,5 +1,0 @@
-'use client'
-
-export default function RootLayout({ children }: { children: React.ReactNode }) {
-    return <>{children}</>
-}
