--- genai-cookbook/app-orig/cookbook/multiturn-chat/recipe.json
@@ -1,4 +1,0 @@
-{
-    "title": "Multi-turn Chat",
-    "description": "This recipe demonstrates a Mantine-powered chat surface that streams multi-turn conversations through the Vercel AI SDK, letting you toggle between Modular MAX and other OpenAI-compatible endpoints without rewriting UI logic. The page component keeps composer input, scroll-follow behavior, and the live message list in sync while forwarding each prompt to a Next.js API route. That route adapts the chat transcript into the SDK’s message format, invokes the selected model via `openai.chat`, and streams tokens back to the browser so replies render fluidly as they arrive."
-}
