--- genai-cookbook/app-orig/page.tsx
@@ -1,6 +1,0 @@
-import { redirect } from 'next/navigation'
-import { cookbookRoute } from '@/lib/constants'
-
-export default function Home() {
-    redirect(cookbookRoute())
-}
