+++ genai-cookbook/app/cookbook/[recipe]/page.tsx
@@ -1,6 +1,10 @@
 import recipeStore from '@/store/RecipeStore'
 
-export default async function Page({ params }: { params: { recipe?: string } }) {
+export default async function RecipeContent({
+    params,
+}: {
+    params: { recipe?: string }
+}) {
     const RecipeComponent = await recipeStore.getComponent(params.recipe)
 
     if (!RecipeComponent) throw new Error(`Recipe Not Found ${params.recipe}`)
