+++ genai-cookbook/theme/globals.scss
@@ -1,4 +1,4 @@
-@use '../lib/partials/_base' as *;
+@use '@/theme/partials/_base' as *;
 
 @tailwind base;
 @tailwind components;
