+++ genai-cookbook/theme/partials/_mantine.scss
